create central authorization system

+++ src/main/java/ru/istu/b1978201/KSite/controllers/ArticleController.java
@@ -48,10 +48,8 @@ public class ArticleController {
     private ArticleService articleService;
 
 
-
-
     @GetMapping("article")
-    public String getArticle(@RequestParam(value = "id",defaultValue = "") String id,Model model) {
+    public String getArticle(@RequestParam(value = "id", defaultValue = "") String id, Model model) {
 
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
 
@@ -60,11 +58,11 @@ public class ArticleController {
         model.addAttribute("user", user);
 
 
-        if(id.isEmpty()) {
+        if (id.isEmpty()) {
             Page<Article> all = articleDao.findAll(PageRequest.of(0, 10, Sort.by(Sort.Order.desc("id"))));
             model.addAttribute("articles", all);
             return "articlelist";
-        }else{
+        } else {
             Article article = articleDao.findByHash(id);
             if (article != null) {
                 model.addAttribute("url", article.getIconUrl());
@@ -75,6 +73,9 @@ public class ArticleController {
             if (article != null) {
                 model.addAttribute("comments", article.getComment());
             }
+            if (article!=null && user != null ) {
+                model.addAttribute("is_owner", article.getUser().getId().equals(user.getId()));
+            }
 
             return "article";
         }
@@ -83,12 +84,23 @@ public class ArticleController {
 
     @PostMapping(value = "article", params = "delete")
     public String deleteArticle(@ModelAttribute("id") String id, Model model) {
-        Article article = articleDao.findByHash(id);
 
-        if (article != null) {
-            commentDao.deleteAll(article.getComment());
-            likeDislikeDao.deleteAll(article.getLikeDislikes());
-            articleDao.delete(article);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        User user = authentication.getPrincipal() instanceof User ? (User) authentication.getPrincipal() : null;
+        model.addAttribute("auth", user != null);
+        model.addAttribute("user", user);
+
+        if (user != null) {
+            Article article = articleDao.findByHash(id);
+            if (article != null) {
+                if (article.getUser().getId().equals(user.getId())) {
+                    commentDao.deleteAll(article.getComment());
+                    likeDislikeDao.deleteAll(article.getLikeDislikes());
+                    articleDao.delete(article);
+                }
+                model.addAttribute("is_owner", article.getUser().getId().equals(user.getId()));
+            }
         }
 
         return "redirect:/article";
@@ -121,13 +133,16 @@ public class ArticleController {
 
 
         if (article != null && user != null && !comment.isEmpty()) {
-            Comment newComment = new Comment();
-            newComment.setComment(comment);
-            newComment.setArticle(article);
-            newComment.setUser(user);
-            article.getComment().add(newComment);
-            commentDao.save(newComment);
-            model.addAttribute("comments", article.getComment());
+            if (article.getUser().getId().equals(user.getId())) {
+                Comment newComment = new Comment();
+                newComment.setComment(comment);
+                newComment.setArticle(article);
+                newComment.setUser(user);
+                article.getComment().add(newComment);
+                commentDao.save(newComment);
+                model.addAttribute("comments", article.getComment());
+            }
+            model.addAttribute("is_owner", article.getUser().getId().equals(user.getId()));
         }
 
         return "article";
@@ -168,7 +183,6 @@ public class ArticleController {
     }
 
 
-
     @PostMapping(value = "article", params = "like")
     public String like(@ModelAttribute("id") String id, @ModelAttribute("comment") String comment, Model model) {
 
@@ -204,7 +218,6 @@ public class ArticleController {
     }
 
 
-
     @GetMapping("editor")
     public String getEditor(Model model) {
 
@@ -231,17 +244,36 @@ public class ArticleController {
         redirectAttributes.addAttribute("auth", user != null);
         redirectAttributes.addAttribute("user", user);
 
-        if (user != null) {
-            Article article = new Article();
-            article.setText(text);
-            article.setDescription(description);
-            article.setHash(UUID.randomUUID().toString());
-            article.setUser(user);
-            article.setIcon(file.getOriginalFilename());
-            article.setDateCreate(new Date());
 
-            articleDao.save(article);
-            storageService.store(file);
+        if (description != null && !description.isEmpty()
+                && text != null && !text.isEmpty()
+                && file != null && !file.isEmpty()
+                && user != null) {
+            String contentType = file.getContentType();
+            if (contentType != null && (contentType.equals("image/jpeg") || contentType.equals("image/png"))) {
+                Article article = new Article();
+                article.setText(text);
+                article.setDescription(description);
+                article.setHash(UUID.randomUUID().toString());
+                article.setUser(user);
+                article.setIcon(file.getOriginalFilename());
+                article.setDateCreate(new Date());
+
+                articleDao.save(article);
+                storageService.store(file);
+            } else {
+                redirectAttributes.addAttribute("file_not_allowed", true);
+            }
+        }
+
+        if (description == null || description.isEmpty()) {
+            redirectAttributes.addAttribute("null_description", true);
+        }
+        if (text == null || text.isEmpty()) {
+            redirectAttributes.addAttribute("null_text", true);
+        }
+        if (file == null || file.isEmpty()) {
+            redirectAttributes.addAttribute("null_file", true);
         }
 
         Page<Article> all = articleDao.findAll(PageRequest.of(0, 4, Sort.by(Sort.Order.desc("id"))));
@@ -256,7 +288,6 @@ public class ArticleController {
 
 
     /**
-     *
      * @param filename
      * @return
      */
